Updated GUI client
